refactor(preferences): dedicated DateFormatField/TimezoneField, dumb SourceInfo
- Add DateFormatField and TimezoneField to the preferences entity, each owning
  its full row: FormField + Combobox, live example (date format), and its own
  resolved source tooltip message
- Reduce SourceInfo to a presentational (i) trigger with a single message prop
- Share the source-resolution wording via a small sourceMessage helper
- PreferencesForm is now a thin orchestrator: fetch + Form + fields + Save
